Make library more C++ friendly

+++ libscpi/src/debug.c
@@ -35,7 +35,6 @@
  */
 
 #include <stdio.h>
-#include <inttypes.h>
 #include "scpi/debug.h"
 
 /**
@@ -48,7 +47,7 @@ bool_t SCPI_DebugCommand(scpi_t * context) {
 	printf("**DEBUG: %s (\"", context->paramlist.cmd->pattern);
 	res = fwrite(context->paramlist.parameters, 1, context->paramlist.length, stdout);
 	(void)res;
-	printf("\" - %" PRIu32 "\r\n", (uint32_t)context->paramlist.length);
+        printf("\" - %lu\r\n", (unsigned long)context->paramlist.length);
 	
 	return TRUE;
 }
